requirements.txt
flask
flask-sqlalchemy
flask_jwt_extended
web3==5.31.1
MarkupSafe==2.0.1
mock
gevent
gunicorn
flask_sslify
PyMySQL
flask_script
Flask-Cors==3.0.10
werkzeug
python-mcs-sdk==0.1.5
python-dotenv~=0.21.0